nix logseq: Add initial support for logseq.
The xdg-desktop portal additions are to help with some electron errors.
There are still others, so I'm gonig to have to focus on getting
electron/wayland working well at some point. For now it works, but
complains a lot.

For config this will need to be a module at some point. With this config
I'll have to select the logseq and make any per graph configs after a
reboot.