nix: Moving a non-ephermeral /home
I'm spending too much time at the moment maintaining an ephemeral /home.
My current plan is to continue using home-manager to manage /home as
much as possible, but not force it on boot. I've modified
btrfs-root-diff to get the files changed in /home as well so that I can
keep track and try to keep drift to a minimum.